Tests are too fast :)

+++ scripts/test-stopper.js
@@ -27,11 +27,21 @@ let timer;
 fs.watch("runner-result", () => {
   clearTimeout(timer);
   timer = setTimeout(() => {
-    let status = +fs.readFileSync("runner-result", "utf8");
-
-    dockerComposeDown();
-    fs.unlinkSync("runner-result");
-
-    process.exit(status);
+    quit();
   }, 500);
 });
+
+quit();
+
+function quit() {
+  let status = +fs.readFileSync("runner-result", "utf8");
+
+  if(isNaN(status)) {
+    return;
+  }
+
+  dockerComposeDown();
+  fs.unlinkSync("runner-result");
+
+  process.exit(status);
+}
